+++ src/components/Fields/RichText/lib/blockquotes.js
@@ -1,45 +1,10 @@
 /* eslint-disable import/prefer-default-export */
 import {
-  Editor,
-  Range,
   Transforms,
   Node,
 } from 'slate';
 import { get } from 'lodash';
-
-const getContainerBlockAtCursor = (editor) => (
-  Editor.above(editor, {
-    match: (n) => Editor.isBlock(editor, n),
-    mode: 'highest',
-  })
-);
-
-const getContainerBlocksAtSelection = (editor) => {
-  const nodes = Editor.nodes(editor);
-
-  const blocks = [];
-
-  // eslint-disable-next-line no-restricted-syntax
-  for (const node of nodes) {
-    // Top level nodes only
-    if (node[1].length === 1) {
-      blocks.push(node);
-    }
-  }
-
-  return blocks;
-};
-
-const getBlocks = (editor) => {
-  const { selection } = editor;
-  const isCollapsed = selection && Range.isCollapsed(selection);
-
-  if (isCollapsed) {
-    return [getContainerBlockAtCursor(editor)];
-  }
-
-  return getContainerBlocksAtSelection(editor);
-};
+import { getBlocks } from './utils';
 
 const toggleBlock = (editor, block) => {
   const type = get(block, [0, 'type']);
